Реализовал функционал с банковскими картами

+++ OnlineBanking.Domain/Entity/Transaction.cs
@@ -12,11 +12,11 @@ namespace OnlineBanking.Domain.Entity
     {
         public long Id { get; set; }
 
-        public int SenderId { get; set; }
+        public long SenderId { get; set; }
 
         public User Sender { get; set; }
 
-        public int RecipientId { get; set; }
+        public long RecipientId { get; set; }
 
         public User Recipient { get; set; }
